Filtrare Dimensiune Maxima
Issue: #10
Facand referinta la Issue #22 si dimensiunea setata anterior, implementam
clasa "FileDimensionFilter" ce are ca scop filtrarea tuturor fisierelor
descarcate prin compararea numarului total de bytes cu dimensiunea maxima
admisa. Aceasta transforma valoarea stocata anterior la Issue #22
"maxDim" in MegaBytes.
Modificand "FileDownloaderImpl" filtram toate fisierele descarcate
conform "FileDimensionFilter". In cazul in care limita admisa este
depasita sau "maxDim" nu este setat conditia devine false si este
aruncata o exceptie IOException cu mesaj sugestiv.

+++ src/ro/mta/facc/webcrawler/download/FileDownloaderImpl.java
@@ -59,7 +59,7 @@ public class FileDownloaderImpl implements FileDownloader {
 
             byte[] dataBuffer = new byte[1024];
             int bytesRead;
-            int totalBytes = 0;
+            long totalBytes = 0;
 
             if (crawlerConfig.getLogLevel() >= 3) {
                 logger.info(String.format("Se descarca %s", downloadUrl));
